Refactor auth feature folder structure
